+++ src/utils/generateTSProperty.ts
@@ -1,9 +1,8 @@
-import {IOpenAPIReferenceObject, IOpenAPISchemaObject, ITypeScriptProperty, SchemaObjectType} from '../types';
-import {getSchemaNameFromRef} from './getSchemaNameFromRef';
-import {getSchemaObjectType} from './getSchemaObjectType';
+import {IOpenAPIReferenceObject, IOpenAPISchemaObject, ITypeScriptProperty} from '../types';
+import {getInterfaceNameFromRef} from './getInterfaceNameFromRef';
 import {isReferenceObject} from './isReferenceObject';
+import {mapSchemaObjectToTypeScriptType} from './mapSchemaObjectToTypeScriptType';
 import {toCamelCase} from './toCamelCase';
-import {toInterfaceName} from './toInterfaceName';
 
 /**
  * Overload for a schemaObject of type IOpenAPIReferenceObject, since we need to compute the name.
@@ -28,7 +27,7 @@ export function generateTSProperty(
  * Converts a OpenAPI 3.0 property to TypeScript property object.
  * @param schemaObject the schema object to convert the properties of.
  * @param requiredProperties an array of required properties for the schema.
- *  @param name the name of the property to be generated.
+ * @param name the name of the property to be generated.
  */
 export function generateTSProperty(
   schemaObject: IOpenAPISchemaObject | IOpenAPIReferenceObject,
@@ -38,45 +37,22 @@ export function generateTSProperty(
   let generatedProperty: ITypeScriptProperty;
 
   if (isReferenceObject(schemaObject)) {
-    const propertyName = toCamelCase(getSchemaNameFromRef(schemaObject.$ref));
+    const propertyName = toCamelCase(getInterfaceNameFromRef(schemaObject.$ref));
     generatedProperty = {
       name: propertyName,
       nullable: false,
       optional: !requiredProperties || !requiredProperties.includes(propertyName),
-      type: ''
+      type: mapSchemaObjectToTypeScriptType(schemaObject)
     };
   } else {
-    const propertyName = name || '';
+    /** Fall back to a generic map. */
+    const propertyName = name || '[key: string]';
     generatedProperty = {
       name: propertyName,
       nullable: schemaObject.nullable || false,
       optional: !requiredProperties || !requiredProperties.includes(propertyName),
-      type: ''
+      type: mapSchemaObjectToTypeScriptType(schemaObject)
     };
-  }
-
-  const schemaObjectType: SchemaObjectType | undefined = getSchemaObjectType(schemaObject);
-
-  switch (schemaObjectType) {
-    case 'ref':
-      generatedProperty.type = toInterfaceName(getSchemaNameFromRef(schemaObject.$ref));
-      break;
-    case 'allOf': // TODO: Support allOf
-    case 'anyOf': // TODO: Support anyOf
-    case 'oneOf': // TODO: Support oneOf
-    case 'enum': // TODO: Support ENUM
-    case 'array': // TODO: Support array
-    case 'string':
-    case 'number':
-    case 'boolean':
-      generatedProperty.type = schemaObjectType;
-      break;
-    case 'object':
-    default:
-      /** Fall back to a generic map. */
-      generatedProperty.name = '[key: string]';
-      generatedProperty.type = 'any';
-      break;
   }
 
   return generatedProperty;
