+++ src/utils/getInterfaceNameFromRef.ts
@@ -1,0 +1,11 @@
+import {toInterfaceName} from './toInterfaceName';
+
+/**
+ * Takes an OpenAPI $ref value and returns the name of the interface.
+ * For example, `#/components/schemas/NewPet` becomes `INewPet`.
+ * @param ref the value of the ref, such as `#/components/schemas/NewPet`.
+ */
+export const getInterfaceNameFromRef = (ref: string | undefined): string => {
+  if (!ref) return '';
+  return toInterfaceName(ref.slice(ref.lastIndexOf('/') + 1));
+};
