+++ src/utils/index.ts
@@ -4,9 +4,9 @@ export * from './convertTSPropertyToString';
 export * from './generateTSInterface';
 export * from './generateTSProperties';
 export * from './generateTSProperty';
+export * from './getInterfaceNameFromRef';
 export * from './getLocalFile';
 export * from './getRemoteFile';
-export * from './getSchemaNameFromRef';
 export * from './getSchemaObjectType';
 export * from './isReferenceObject';
 export * from './toCamelCase';
